Rename file and add Spring dependecies for HTTP request mock
