--- src/test/CreatePlaythroughServletTest.java
@@ -1,45 +1,0 @@
-// Copyright 2019 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.google.sps;
-
-import com.google.sps.servlets.CreateRoomServlet;
-
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.JUnit4;
-
-@RunWith(JUnit4.class)
-public static final class CreatePlaythroughServletTest {
-
-  @Mock private RoomDao mockRoomDao;
-  @Mock private ServletContext mockServletContext;
-
-  private CreateRoomServlet servletUnderTest;
-
-  @Before
-  public void setUp() throws Exception {
-    servletUnderTest = new CreateRoomServlet() {
-      @Override
-      public ServletContext() {
-        return mockServletContext;
-      }
-    };
-
-    when(mockServletContext.getAttribute("roomDao")).thenReturn(mockRoomDao);
-  }
-  
-}
